refactor dept 패키지 생성 하면 오류 발생 -> 다시 원위치로
