--- src/main/java/com/ideality/coreflow/dept/command/application/service/DeptService.java
@@ -1,4 +1,0 @@
-package com.ideality.coreflow.dept.command.application.service;
-
-public interface DeptService {
-}
