+++ src/main/java/com/ideality/coreflow/project/command/application/service/DeptService.java
@@ -1,0 +1,4 @@
+package com.ideality.coreflow.project.command.application.service;
+
+public interface DeptService {
+}
