+++ src/main/java/com/ideality/coreflow/project/command/application/service/impl/WorkDeptServiceImpl.java
@@ -2,8 +2,6 @@ package com.ideality.coreflow.project.command.application.service.impl;
 
 import com.ideality.coreflow.project.command.application.service.WorkDeptService;
 import com.ideality.coreflow.project.command.domain.aggregate.WorkDept;
-import com.ideality.coreflow.dept.command.domain.repository.DeptRepository;
-import com.ideality.coreflow.project.command.domain.repository.TaskRepository;
 import com.ideality.coreflow.project.command.domain.repository.WorkDeptRepository;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
@@ -16,8 +14,6 @@ import org.springframework.transaction.annotation.Transactional;
 public class WorkDeptServiceImpl implements WorkDeptService {
 
     private final WorkDeptRepository workDeptRepository;
-    private final TaskRepository taskRepository;
-    private final DeptRepository deptRepository;
 
     @Override
     @Transactional
