+++ src/main/java/com/ideality/coreflow/project/command/domain/repository/DeptRepository.java
@@ -1,6 +1,6 @@
-package com.ideality.coreflow.dept.command.domain.repository;
+package com.ideality.coreflow.project.command.domain.repository;
 
-import com.ideality.coreflow.dept.command.domain.aggregate.Dept;
+import com.ideality.coreflow.project.command.domain.aggregate.Dept;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
