+++ src/main/resources/mapper/DeptMapper.xml
@@ -2,7 +2,7 @@
 <!DOCTYPE mapper
         PUBLIC "-//mybatis.org//DTD Mapper 3.0//EN"
         "https://mybatis.org/dtd/mybatis-3-mapper.dtd">
-<mapper namespace="com.ideality.coreflow.dept.query.mapper.DeptMapper">
+<mapper namespace="com.ideality.coreflow.project.query.mapper.DeptMapper">
     <select id="findNameById" resultType="string">
         SELECT name
           FROM dept
