+++ src/test/java/com/ideality/coreflow/project/command/application/controller/TaskControllerTest.java
@@ -5,8 +5,11 @@ import com.ideality.coreflow.project.command.application.dto.RequestTaskDTO;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.boot.autoconfigure.domain.EntityScan;
 import org.springframework.boot.test.autoconfigure.web.servlet.AutoConfigureMockMvc;
 import org.springframework.boot.test.context.SpringBootTest;
+import org.springframework.context.annotation.ComponentScan;
+import org.springframework.data.jpa.repository.config.EnableJpaRepositories;
 import org.springframework.http.MediaType;
 import org.springframework.test.web.servlet.MockMvc;
 
@@ -19,6 +22,9 @@ import static org.springframework.test.web.servlet.result.MockMvcResultMatchers.
 
 @SpringBootTest
 @AutoConfigureMockMvc
+@EntityScan(basePackages = "com.ideality.coreflow")
+@EnableJpaRepositories(basePackages = "com.ideality.coreflow")
+@ComponentScan(basePackages = "com.ideality.coreflow")
 public class TaskControllerTest {
 
     @Autowired
